third_party/blink/web_tests/http/tests/security/location-prototype-expected.txt

ALERT: Yay! Calling a function that shouldn't exist threw an exception.
This test passes if it doesn't alert the string "fail". 
